Separate parser for commands related to contact
